Move db statements to own file
